Move memberships to context
